// Project/dff_library/ui_components/components/DisclaimerInformation/utils/enum.ts
export enum DisclaimerEnumValues {
  CustomIconMulti = 'Custom icon (Multiple)',
  Default = 'Default',
  DefaultOrange = 'Default orange',
  GeneralInformation = 'General information',
  Neutral = 'Neutral',
  PositiveInline = 'Positive-Inline',
  PositiveLarge = 'Positive-Large',
  PositiveSmall = 'Positive-Small',
  PreConfirmationSingle = 'Pre-confirmation (single)',
  TipsLarge = 'Tips-Large',
  TipsSmall = 'Tips-Small',
  Warning = 'Warning',
}